fixed some static analysis warnings

+++ src/org/acplt/oncrpc/apps/jrpcgen/cup_runtime/virtual_parse_stack.java
@@ -85,7 +85,7 @@ public class virtual_parse_stack {
         real_next++;
 
         /* put the state number from the Symbol onto the virtual stack */
-        vstack.push(Integer.valueOf(stack_sym.parse_state));
+        vstack.push(stack_sym.parse_state);
     }
 
     /*. . . . . . . . . . . . . . . . . . . . . . . . . . . . . .*/
@@ -104,7 +104,7 @@ public class virtual_parse_stack {
                     "Internal parser error: top() called on empty virtual stack");
         }
 
-        return vstack.peek().intValue();
+        return vstack.peek();
     }
 
     /*. . . . . . . . . . . . . . . . . . . . . . . . . . . . . .*/
@@ -127,7 +127,7 @@ public class virtual_parse_stack {
     /*. . . . . . . . . . . . . . . . . . . . . . . . . . . . . .*/
     /** Push a state number onto the stack. */
     public void push(int state_num) {
-        vstack.push(Integer.valueOf(state_num));
+        vstack.push(state_num);
     }
 
     /*-----------------------------------------------------------*/
